requirements.txt
transformers==4.45.2  # Pinned version for compatibility
datasets>=2.12.0
accelerate>=0.20.0
torch>=2.0.0
evaluate>=0.4.0
jiwer>=3.0.0
soundfile>=0.12.1
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
librosa>=0.10.1
pyyaml>=6.0.1
flash-attn>=2.7.2  # Required for optimized attention computation 